Fix part of Google system tests (#18494)
* Fix CloudBuildExampleDagsSystemTest

* Fix DataCatalog tests

* Fix Dataprep test

* Fix GCS test

* Fix Dataflow test

* Fix GC Functions test

* Fix Cloud SQL tests

* Fix Data Fusion test

* Change env var in Dataprep system test

* Fix config for dataflow system tests

* Add unit test for datafusion

* Fix Data Fusion utest, add optional parameter to data fusion pipeline state sensor

* Update code with black suggestions

* Fix flake8 issue

* Revert changes to example_cloud_sql.py according to review feedback

+++ tests/providers/google/cloud/operators/test_gcs_system_helper.py
@@ -52,9 +52,12 @@ with open(source, "r") as src, open(destination, "w+") as dest:
 
     @staticmethod
     def remove_test_files():
-        os.remove(PATH_TO_UPLOAD_FILE)
-        os.remove(PATH_TO_SAVED_FILE)
-        os.remove(PATH_TO_TRANSFORM_SCRIPT)
+        if os.path.exists(PATH_TO_UPLOAD_FILE):
+            os.remove(PATH_TO_UPLOAD_FILE)
+        if os.path.exists(PATH_TO_SAVED_FILE):
+            os.remove(PATH_TO_SAVED_FILE)
+        if os.path.exists(PATH_TO_TRANSFORM_SCRIPT):
+            os.remove(PATH_TO_TRANSFORM_SCRIPT)
 
     def remove_bucket(self):
         self.execute_cmd(["gsutil", "rm", "-r", f"gs://{BUCKET_1}"])
